orchestrator/pyproject.toml
[project]
name = "eval-orchestrator"
version = "0.1.0"
description = "Agentic evaluation system orchestrator for testing model/harness combinations"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "litellm>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.1",
    "pre-commit>=3.0",
]

[project.scripts]
eval-orchestrator = "agentic_eval.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentic_eval"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "C90"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
